+++ src/main/java/Camioneta.java
@@ -3,7 +3,7 @@ package vehiculos;
 public class Camioneta extends Vehiculo{
 
 	private boolean volco;
-	public Camioneta( String placa, int puertas, String nombre, double precio, double peso, Fabricante fabricante, boolean volco) {
+	public Camioneta( String placa, int puertas, String nombre, int precio, int peso, Fabricante fabricante, boolean volco) {
 
 		super(placa, puertas, nombre, precio, peso, fabricante);
 		this.volco=volco;
